+++ src/extension.ts
@@ -10,10 +10,12 @@ export function activate(context: vscode.ExtensionContext) {
 		vscode.commands.registerCommand('sfk.studio.goToDefinitionConfiguration', GoToDefinitionConfiguration.execute),
 		vscode.languages.registerDefinitionProvider('typescript', { provideDefinition: GoToDefinitionTypeScript.execute }),
 		vscode.languages.registerImplementationProvider('typescript', { provideImplementation: GoToDefinitionTypeScript.execute }),
-		vscode.workspace.onDidSaveTextDocument(CompileScript.executeTranspile),
+		vscode.workspace.onDidSaveTextDocument(CompileScript.executeCompile),
 		vscode.workspace.onDidSaveTextDocument(CompileScript.executeLibrariesDeclaration),
 		vscode.workspace.onDidRenameFiles(CompileScript.executeRename)
 	);
+
+	CompileScript.executeCompileAll();
 }
 
 export function deactivate() {}