+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "sfk-studio" extension will be documented in this fil
 ### Added
 - [TypeScript]: Compile with SFK transform on saving
 - [TypeScript]: Generate Script Libraries Declaration files
+- [TypeScript]: Compile all on extension activation
 
 ### Fixed
 - [TypeScript]: Rename/Move compiled JS file on TS file renaming/moving